refactor: 옵션 생성, 추가 옵션 상품 생성 리팩토링
