+++ src/components/productOption/AdditionalOptionListItem.js
@@ -1,10 +1,7 @@
-import { productOptionState } from "atoms/productOption";
 import CustomButton from "components/common/CustomButton";
 import Input from "components/common/Input";
 import useProductOption from "hooks/useProductOption";
-import produce from "immer";
 import React from "react";
-import { useSetRecoilState } from "recoil";
 import styled from "styled-components";
 
 function AdditionalOptionListItem({
