+++ src/hooks/useProductOption.js
@@ -8,9 +8,7 @@ import { v4 as uuidv4 } from "uuid";
 export default function useProductOption() {
   const [productOption, setProductOption] = useRecoilState(productOptionState);
 
-  /**
-   * @returns 옵션 세트 추가
-   */
+  // @Note 옵션 세트 추가
   const onCreateOptionSet = useCallback(() => {
     setProductOption((prevOptions) => [
       ...prevOptions,
@@ -32,12 +30,87 @@ export default function useProductOption() {
     ]);
   }, [setProductOption]);
 
-  /**
-   * @param {string} optionSetId 옵션 세트 id
-   * @param {string} optionId 옵션 id
-   * @param {string} additionalOptionId 추가 옵션 id
-   * @returns 추가 옵션 상품 삭제
-   */
+  // @Note 옵션 추가
+  const onCreateOption = useCallback(
+    (optionSetId) => {
+      setProductOption((prevOptions) =>
+        produce(prevOptions, (draft) => {
+          for (let i = 0; i < draft.length; i++) {
+            if (draft[i].id === optionSetId) {
+              draft[i].options.push({
+                id: uuidv4(),
+                name: "",
+                normalPrice: 0,
+                salePrice: 0,
+                stock: 0,
+                tax: "",
+                additionalOptions: [],
+              });
+              break;
+            }
+          }
+        })
+      );
+    },
+    [setProductOption]
+  );
+
+  // @Note 옵션 삭제 (옵션이 1개면 해당 옵션 세트 자체를 지워주기)
+  const onDeleteOption = useCallback(
+    (optionSetId, optionId) => {
+      const optionLength = productOption.filter(
+        (option) => option.id === optionSetId
+      )[0].options.length;
+
+      if (optionLength === 1) {
+        setProductOption((prevOptions) =>
+          prevOptions.filter(({ id }) => id !== optionSetId)
+        );
+        return;
+      }
+
+      setProductOption((prevOptions) =>
+        produce(prevOptions, (draft) => {
+          for (let i = 0; i < draft.length; i++) {
+            if (draft[i].id === optionSetId) {
+              draft[i].options = draft[i].options.filter(
+                (item) => item.id !== optionId
+              );
+              break;
+            }
+          }
+        })
+      );
+    },
+    [productOption, setProductOption]
+  );
+
+  // @Note 추가 옵션 생성
+  const onCreateAdditionalOption = useCallback(
+    (optionSetId, optionId) => {
+      setProductOption((prevOptions) =>
+        produce(prevOptions, (draft) => {
+          for (let i = 0; i < draft.length; i++) {
+            if (draft[i].id === optionSetId) {
+              draft[i].options.map((item) => {
+                if (item.id === optionId) {
+                  item.additionalOptions.push({
+                    id: uuidv4(),
+                    additionalName: "",
+                    additionalNormalPrice: 0,
+                  });
+                }
+              });
+              break;
+            }
+          }
+        })
+      );
+    },
+    [setProductOption]
+  );
+
+  // @Note 추가 옵션 상품 삭제
   const onDeleteAdditionalOption = useCallback(
     (optionSetId, optionId, additionalOptionId) => {
       setProductOption((prevOptions) =>
@@ -58,11 +131,14 @@ export default function useProductOption() {
         })
       );
     },
-    []
+    [setProductOption]
   );
 
   return {
     onCreateOptionSet,
+    onCreateOption,
+    onDeleteOption,
+    onCreateAdditionalOption,
     onDeleteAdditionalOption,
   };
 }
